requirements.txt
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
PyYAML>=6.0
pyarrow>=12.0.0
# MetaTrader5>=5.0.45  # Windows only — uncomment on inference machine
